requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
APScheduler==3.10.4
attrs==23.2.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pytz==2024.1
six==1.16.0
tzdata==2024.1
tzlocal==5.2
yarl==1.9.4
